feat: refactor middlewares and update package dependencies
