+++ examples/hello-world/app/example/index.mdx
@@ -1,3 +1,5 @@
+export const year = 2023
+
 # Last year’s snowfall
 
 In {year}, the snowfall was above average.
