+++ examples/hello-world/app/example/index.tsx
@@ -1,5 +1,12 @@
 import { mdx } from 'buntal'
 
-export default function Example() {
-  return <div>{mdx('./index.mdx')}</div>
+export const $ = () => mdx('./app/example/index.mdx')
+
+export default function Example({
+  data
+}: {
+  data?: Awaited<ReturnType<typeof $>>
+}) {
+  // eslint-disable-next-line @eslint-react/dom/no-dangerously-set-innerhtml
+  return <div dangerouslySetInnerHTML={{ __html: data || '' }}></div>
 }
