+++ examples/hello-world/app/example/layout.tsx
@@ -1,0 +1,3 @@
+export default function MDLayout({ children }: { children: React.ReactNode }) {
+  return <div className="container mx-auto prose">{children}</div>
+}
