+++ packages/@buntal/http/middlewares/index.ts
@@ -1,3 +1,3 @@
 export * from './auth'
-export * from './basic'
 export * from './cors'
+export * from './logger'
