+++ packages/@buntal/http/router/index.ts
@@ -1,5 +1,5 @@
 export const buildRouter = (dir: string) =>
   new Bun.FileSystemRouter({
     style: 'nextjs',
-    dir: dir
+    dir
   })
