--- packages/@buntal/middlewares/basic/index.ts
@@ -1,1 +1,0 @@
-export { logger } from './logger'
