--- packages/@buntal/middlewares/package.json
@@ -1,16 +1,0 @@
-{
-  "name": "@buntal/middlewares",
-  "version": "0.0.5",
-  "module": "index.ts",
-  "type": "module",
-  "devDependencies": {
-    "@types/bun": "latest"
-  },
-  "peerDependencies": {
-    "typescript": "^5"
-  },
-  "dependencies": {
-    "@buntal/http": "latest",
-    "jose": "^6.0.11"
-  }
-}
