+++ packages/buntal/package.json
@@ -20,7 +20,6 @@
   },
   "dependencies": {
     "@buntal/http": "workspace:^",
-    "@buntal/middlewares": "workspace:^",
     "@mdx-js/mdx": "^3.1.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0"
